+++ lib/shoulda/matchers/active_model/numericality_matchers/odd_number_matcher.rb
@@ -4,15 +4,7 @@ module Shoulda
       module NumericalityMatchers
         # @private
         class OddNumberMatcher < NumericTypeMatcher
-          NON_ODD_NUMBER_VALUE  = 2
-
-          def initialize(attribute, options = {})
-            @attribute = attribute
-            @disallow_value_matcher =
-              DisallowValueMatcher.new(NON_ODD_NUMBER_VALUE.to_s).
-              for(@attribute).
-              with_message(:odd)
-          end
+          NON_ODD_NUMBER_VALUE = 2
 
           def allowed_type
             'odd numbers'
@@ -20,6 +12,20 @@ module Shoulda
 
           def diff_to_compare
             2
+          end
+
+          protected
+
+          def wrap_disallow_value_matcher(matcher)
+            matcher.with_message(:odd)
+          end
+
+          def disallowed_value
+            if @numeric_type_matcher.given_numeric_column?
+              NON_ODD_NUMBER_VALUE
+            else
+              NON_ODD_NUMBER_VALUE.to_s
+            end
           end
         end
       end