+++ lib/shoulda/matchers/active_model/numericality_matchers/only_integer_matcher.rb
@@ -5,13 +5,6 @@ module Shoulda
         # @private
         class OnlyIntegerMatcher < NumericTypeMatcher
           NON_INTEGER_VALUE = 0.1
-          def initialize(attribute)
-            @attribute = attribute
-            @disallow_value_matcher =
-              DisallowValueMatcher.new(NON_INTEGER_VALUE.to_s).
-              for(attribute).
-              with_message(:not_an_integer)
-          end
 
           def allowed_type
             'integers'
@@ -19,6 +12,20 @@ module Shoulda
 
           def diff_to_compare
             1
+          end
+
+          protected
+
+          def wrap_disallow_value_matcher(matcher)
+            matcher.with_message(:not_an_integer)
+          end
+
+          def disallowed_value
+            if @numeric_type_matcher.given_numeric_column?
+              NON_INTEGER_VALUE
+            else
+              NON_INTEGER_VALUE.to_s
+            end
           end
         end
       end